requirements/local.txt
-r base.txt

pre-commit==3.6.0
django-extensions==3.2.3
Werkzeug==3.0.1
coverage==7.3.4
freezegun==1.4.0
black==24.2.0
flake8==7.0.0
isort==5.13.2
ruff==0.2.2
